Move display name setter to shared
