+++ packages/react/src/ReactForwardRef.js
@@ -6,6 +6,7 @@
  */
 
 import {REACT_FORWARD_REF_TYPE, REACT_MEMO_TYPE} from 'shared/ReactSymbols';
+import setDisplayName from 'shared/setDisplayName';
 
 export function forwardRef<Props, ElementType: React$ElementType>(
   render: (props: Props, ref: React$Ref<ElementType>) => React$Node,
@@ -48,28 +49,7 @@ export function forwardRef<Props, ElementType: React$ElementType>(
     render,
   };
   if (__DEV__) {
-    let ownName;
-    Object.defineProperty(elementType, 'displayName', {
-      enumerable: false,
-      configurable: true,
-      get: function() {
-        return ownName;
-      },
-      set: function(name) {
-        ownName = name;
-
-        // The inner component shouldn't inherit this display name in most cases,
-        // because the component may be used elsewhere.
-        // But it's nice for anonymous functions to inherit the name,
-        // so that our component-stack generation logic will display their frames.
-        // An anonymous function generally suggests a pattern like:
-        //   React.forwardRef((props, ref) => {...});
-        // This kind of inner function is not used elsewhere so the side effect is okay.
-        if (!render.name && !render.displayName) {
-          render.displayName = name;
-        }
-      },
-    });
+    setDisplayName(elementType, render)
   }
   return elementType;
 }
