+++ packages/react/src/ReactMemo.js
@@ -8,6 +8,7 @@
 import {REACT_MEMO_TYPE} from 'shared/ReactSymbols';
 
 import isValidElementType from 'shared/isValidElementType';
+import setDisplayName from 'shared/setDisplayName';
 
 export function memo<Props>(
   type: React$ElementType,
@@ -28,28 +29,7 @@ export function memo<Props>(
     compare: compare === undefined ? null : compare,
   };
   if (__DEV__) {
-    let ownName;
-    Object.defineProperty(elementType, 'displayName', {
-      enumerable: false,
-      configurable: true,
-      get: function() {
-        return ownName;
-      },
-      set: function(name) {
-        ownName = name;
-
-        // The inner component shouldn't inherit this display name in most cases,
-        // because the component may be used elsewhere.
-        // But it's nice for anonymous functions to inherit the name,
-        // so that our component-stack generation logic will display their frames.
-        // An anonymous function generally suggests a pattern like:
-        //   React.memo((props) => {...});
-        // This kind of inner function is not used elsewhere so the side effect is okay.
-        if (!type.name && !type.displayName) {
-          type.displayName = name;
-        }
-      },
-    });
+    setDisplayName(elementType, type)
   }
   return elementType;
 }
